07-28
加款功能完善
绘制二级代理商给下级代理商添加折扣页面

+++ weizu-parent/weizu-web-modules/weizu-channel/src/main/java/com/weizu/flowsys/web/agency/dao/impl/ChargeAccountDao.java
@@ -13,7 +13,7 @@ import com.weizu.flowsys.web.agency.dao.ChargeAccountDaoInterface;
 import com.weizu.flowsys.web.agency.pojo.ChargeAccountPo;
 
 @Repository("chargeAccountDao")
-public class ChargeAccountDao extends DaoImpl<ChargeAccountPo, Long> implements ChargeAccountDaoInterface {
+public class ChargeAccountDao extends DaoImpl<ChargeAccountPo, Integer> implements ChargeAccountDaoInterface {
 
 	@Resource
 	private SqlSessionTemplate sqlSessionTemplate;
